+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotImpl.java
@@ -58,6 +58,7 @@ import java.util.logging.Handler;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Engine;
 import org.graalvm.polyglot.HostAccess;
+import org.graalvm.polyglot.HostAccess.TargetMappingPrecedence;
 import org.graalvm.polyglot.PolyglotException;
 import org.graalvm.polyglot.ResourceLimitEvent;
 import org.graalvm.polyglot.Value;
@@ -284,9 +285,9 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
     }
 
     @Override
-    public <S, T> Object newTargetTypeMapping(Class<S> sourceType, Class<T> targetType, Predicate<S> acceptsValue, Function<S, T> convertValue) {
+    public <S, T> Object newTargetTypeMapping(Class<S> sourceType, Class<T> targetType, Predicate<S> acceptsValue, Function<S, T> convertValue, TargetMappingPrecedence precedence) {
         try {
-            return new PolyglotTargetMapping(sourceType, targetType, acceptsValue, convertValue);
+            return new PolyglotTargetMapping(sourceType, targetType, acceptsValue, convertValue, precedence);
         } catch (Throwable t) {
             throw PolyglotImpl.guestToHostException(this, t);
         }